# fix(vm): Release VM properly