--- a/gnovm/pkg/gnolang/machine_test.go
+++ b/gnovm/pkg/gnolang/machine_test.go
@@ -1,0 +1,10 @@
+package gnolang
+
+import "testing"
+
+func BenchmarkCreateNewMachine(b *testing.B) {
+	for i := 0; i < b.N; i++ {
+		m := NewMachineWithOptions(MachineOptions{})
+		m.Release()
+	}
+}