@ajnavarro 2023-09-19
@thehowl  Applied requested changes and added benchmarks:

```
PR with makes:
BenchmarkCreateNewMachine-12    	  160599	     13276 ns/op	   45496 B/op	      35 allocs/op

Master branch:
BenchmarkCreateNewMachine-12    	  317624	      7059 ns/op	    3243 B/op	      32 allocs/op

Add changes to the release method:
BenchmarkCreateNewMachine-12    	  318438	      7214 ns/op	    3246 B/op	      32 allocs/op

Add changes to the NewMachineWithOptions method:
BenchmarkCreateNewMachine-12    	  323756	      7193 ns/op	    3240 B/op	      32 allocs/op
```